Move fork time graph out to separate tab.
