--- public/js/components/Charts/Charts.js
@@ -1,98 +1,0 @@
-import React from 'react';
-import { Themes } from 'formidable-charts';
-import { Grid, Row, Col } from 'react-flexbox-grid';
-import AreaChartWrap from './AreaChartWrap';
-import ScatterChartWrap from './ScatterChartWrap';
-import { CHART_FILTERS_MAP } from "../../utils/chartList";
-
-const NoDataMessage = ({ filters }) =>
-    <div className="no-data">
-        No data, please update any of the following:<br />
-        <span className="filters">{filters.join(", ")}</span>
-    </div>
-
-const Charts = ({ charts, isUpdating, toggleStackChart, filterVals }) => {
-
-    const renderTitle = (type) => {
-        switch (type) {
-        case 'originatingSystem':
-            return <h3>Tool of Origin: <span className='chart-title--second-series'>InCopy</span> vs <span className='chart-title--first-series'>Composer</span></h3>;
-        case 'workflowState':
-            return <h3>Workflow State: <span className='chart-title--second-series'>Never Seen in Workflow</span> vs <span className='chart-title--first-series'>Seen in Workflow</span></h3>;
-        case 'forkTime':
-            return <h3>Fork Time</h3>;
-        default:
-            '';
-        }
-    };
-
-    const getClassName = (error) => {
-        if (error) {
-            return 'chart-wrap chart-wrap__error';
-        } else {
-            return isUpdating ? 'chart-wrap chart-wrap__updating' : 'chart-wrap';
-        }
-    };
-
-    return (
-        <Grid fluid>
-            <Row around='xs'>
-                <Col xs={12} md={6}>
-                    <AreaChartWrap
-                        getClassName={getClassName}
-                        filterVals={filterVals}
-                        hasCsvButton={true}
-                        toggleStackChart={toggleStackChart}
-                        chartType={'ComposerVsIncopy'}
-                        titleHeader={renderTitle('originatingSystem')}
-                        scale='time'
-                        data={charts.composerVsInCopy.data}
-                        isStacked={charts.composerVsInCopy.isStacked}
-                        error={charts.composerVsInCopy.error}
-                        hasToggle={true}
-                        height={250}
-                        yLabel='Published Content Daily Numbers'
-                        yLabelStacked='Published Content Daily %'
-                    />
-                </Col>
-                <Col xs={12} md={6}>
-                    <AreaChartWrap
-                        getClassName={getClassName}
-                        filterVals={filterVals}
-                        hasCsvButton={true}
-                        chartType={'InWorkflowVsNotInWorkflow'}
-                        titleHeader={renderTitle('workflowState')}
-                        scale='time'
-                        data={charts.inWorkflowVsNotInWorkflow.data}
-                        isStacked={charts.inWorkflowVsNotInWorkflow.isStacked}
-                        hasToggle={false}
-                        error={charts.inWorkflowVsNotInWorkflow.error}
-                        height={250}
-                        yLabel='Published Content Daily Numbers'
-                        yLabelStacked='Published Content Daily %'
-                    />
-                </Col>
-            </Row>
-            <Row around='xs'>
-                <Col xs={12} md={8}>
-                    <ScatterChartWrap
-                        hasToggle={false}
-                        error={charts.forkTime.error}
-                        chartType={'ForkTime'}
-                        getClassName={getClassName}
-                        filterVals={filterVals}
-                        titleHeader={renderTitle('forkTime')}
-                        hasCsvButton={true}
-                        scale='time'
-                        data={charts.forkTime.data}
-                        noDataMessage={<NoDataMessage filters={CHART_FILTERS_MAP["ForkTime"]} />}
-                        height={250}
-                        yLabel='Time to Publication'
-                    />
-                </Col>
-            </Row>
-        </Grid>
-    );
-};
-
-export default Charts;
